backend/requirements.txt
fastapi
sqlalchemy
sqlmodel
asyncpg
uvicorn
python-dotenv
pydantic-settings
